move notes functions to their own file to prevent circular dependencies
