+++ app/modules/kit/service.server.ts
@@ -33,10 +33,8 @@ import type { MergeInclude } from "~/utils/utils";
 import type { UpdateKitPayload } from "./types";
 import { GET_KIT_STATIC_INCLUDES, KITS_INCLUDE_FIELDS } from "./types";
 import { getKitsWhereInput } from "./utils.server";
-//@TODO: Fix the circular dependency
-// eslint-disable-next-line import/no-cycle
-import { createNote } from "../asset/service.server";
 import type { CreateAssetFromContentImportPayload } from "../asset/types";
+import { createNote } from "../note/service.server";
 import { getQr } from "../qr/service.server";
 
 import { getUserByID } from "../user/service.server";
