+++ app/routes/_layout+/assets.$assetId.overview.release-custody.tsx
@@ -6,8 +6,8 @@ import { Form } from "~/components/custom-form";
 import { UserXIcon } from "~/components/icons/library";
 import { Button } from "~/components/shared/button";
 import { db } from "~/database/db.server";
-import { createNote } from "~/modules/asset/service.server";
 import { releaseCustody } from "~/modules/custody/service.server";
+import { createNote } from "~/modules/note/service.server";
 import { getUserByID } from "~/modules/user/service.server";
 import styles from "~/styles/layout/custom-modal.css?url";
 import { sendNotification } from "~/utils/emitter/send-notification.server";
