+++ app/routes/_layout+/bookings.$bookingId.add-assets.tsx
@@ -44,16 +44,14 @@ import {
 } from "~/components/shared/tabs";
 import { Td } from "~/components/table";
 
-import {
-  createNotes,
-  getPaginatedAndFilterableAssets,
-} from "~/modules/asset/service.server";
+import { getPaginatedAndFilterableAssets } from "~/modules/asset/service.server";
 import {
   getBooking,
   getKitIdsByAssets,
   removeAssets,
   upsertBooking,
 } from "~/modules/booking/service.server";
+import { createNotes } from "~/modules/note/service.server";
 import { getUserByID } from "~/modules/user/service.server";
 import { getClientHint } from "~/utils/client-hints";
 import { makeShelfError } from "~/utils/error";
